#170 Moved test utils
