+++ .github/workflows/ci.yml
@@ -38,7 +38,13 @@ jobs:
         working-directory: ../integration-test-docker-environment
 
       - name: Run Unit Tests
-        run: poetry run pytest test/unit
+        run: poetry run nox -s test:unit
+
+      - name: Upload Artifacts
+        uses: actions/upload-artifact@v3
+        with:
+          name: .coverage
+          path: .coverage
 
   metrics:
     needs: [ fast-tests ]
